+++ velib_modules/model/evaluation.py
@@ -4,8 +4,8 @@ from velib_modules.model.metrics import root_mean_squared_error, mean_percentage
 
 def evaluate_model(model, features_test, y_test):
     y_pred = model.predict(features_test)
-    bikes_pred = y_pred * features_test.bike_stands
-    bikes_actual = y_test * features_test.bike_stands
+    bikes_pred = (y_pred * features_test.bike_stands).astype(int)
+    bikes_actual = (y_test * features_test.bike_stands).astype(int)
     performance = {}
     performance['RMSE'] = root_mean_squared_error(bikes_actual, bikes_pred)
     performance['MAPE'] = mean_percentage_error(bikes_actual, bikes_pred)
